fix: build warning

+++ src/DatabaseManager.cpp
@@ -199,7 +199,7 @@ namespace ModChatTransmitter
     bool DatabaseManager::IsValidUtf8(const Binary& data)
     {
         int cnt = 0;
-        for (int i = 0; i < data.size(); i++)
+        for (size_t i = 0; i < data.size(); i++)
         {
             int x = data[i];
             if (!cnt)
